+++ commands/delete-azure-image.ts
@@ -8,26 +8,17 @@ import { AzureImage, getSub, Repository } from "../explorer/utils/azureUtils";
 import { AzureAccount } from "../typings/azure-account.api";
 import { AzureCredentialsManager } from '../utils/azureCredentialsManager';
 const teleCmdId: string = 'vscode-docker.deleteAzureImage';
+import * as quickPicks from '../commands/utils/quick-pick-azure';
+import * as acrTools from '../utils/Azure/acrTools';
 
 /**
  * function to delete an Azure repository and its associated images
  * @param context : if called through right click on AzureRepositoryNode, the node object will be passed in. See azureRegistryNodes.ts for more info
  */
 export async function deleteAzureImage(context?: AzureImageNode): Promise<void> {
-    let azureAccount: AzureAccount;
-    if (context) {
-        azureAccount = context.azureAccount;
-    } else {
-        azureAccount = await AzureCredentialsManager.getInstance().getAccount();
+    if (!AzureCredentialsManager.getInstance().isLoggedIn()) {
+        vscode.window.showErrorMessage('You are not logged into Azure');
     }
-    if (!azureAccount) {
-        return;
-    }
-
-    if (azureAccount.status === 'LoggedOut') {
-        return;
-    }
-
     let registry: Registry;
     let subscription: SubscriptionModels.Subscription;
     let repoName: string;
@@ -35,11 +26,11 @@ export async function deleteAzureImage(context?: AzureImageNode): Promise<void> 
     let password: string;
     let tag: string;
     if (!context) {
-        registry = await AzureCredentialsManager.getInstance().getRegistry();
+        registry = await quickPicks.quickPickACRRegistry();
         subscription = getSub(registry);
-        let repository: Repository = await AzureCredentialsManager.getInstance().getRepository(registry);
+        let repository: Repository = await quickPicks.quickPickACRRepository(registry);
         repoName = repository.name;
-        const image = await AzureCredentialsManager.getInstance().getImage(repository);
+        const image = await quickPicks.quickPickACRImage(repository);
         tag = image.tag;
     }
 
@@ -48,26 +39,23 @@ export async function deleteAzureImage(context?: AzureImageNode): Promise<void> 
         ignoreFocusOut: true,
         placeHolder: 'No',
         value: 'No',
-        prompt: 'Are you sure you want to delete this image? Enter Yes or No: '
+        prompt: 'Are you sure you want to delete this image? Enter Yes to continue: '
     };
     let answer = await vscode.window.showInputBox(opt);
     if (answer !== 'Yes') { return; }
 
     if (context) {
-        username = context.userName;
-        password = context.password;
         repoName = context.label;
         subscription = context.subscription;
         registry = context.registry;
         let wholeName = repoName.split(':');
         repoName = wholeName[0];
         tag = wholeName[1];
-    } else { //this is separated from !context above so it only calls loginCredentials once user has assured they want to delete the image
-        let creds = await AzureCredentialsManager.getInstance().loginCredentials(subscription, registry);
-        username = creds.username;
-        password = creds.password;
     }
 
+    let creds = await acrTools.loginCredentials(subscription, registry);
+    username = creds.username;
+    password = creds.password;
     let path = `/v2/_acr/${repoName}/tags/${tag}`;
     await azureUtils.request_data_from_registry('delete', registry.loginServer, path, username, password); //official call to delete the image
 }
